Allow specifying build_file in initialize request
Specifying project-specific settings on the command presents a challenge
for editors where configuration such as "how to start the server" is
static. The LSP-way to provide project-specific options is via the
'initializationOptions' in initialize request.

Well, technically, it's via the 'didChangeConfigurationRequest', but the
subtle distinction between these two is mosly ignored in my experience.

ANyway, this allows YCM to provide project-specific value for the build
file, which necessarily differs by project. Ultimately, we probably want
heuristics for this (e.g. serach the filesystem from the root dir
provided on initialize for a build.jai or first.jai), but this is a good
start.